% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CCI_icd10}
\alias{CCI_icd10}
\title{CCI for ICD-10-CM (beta version)}
\format{
\describe{
\item{I10_DX}{ICD-10-CM diagnosis code, without decimials}
\item{I10_DX_desc}{Description of the ICD-10-CM diagnosis code}
\item{CCI}{Categorical variable inidcating if the ICD code represents an
acute or chronic condition (see below)}
Starting in v2021.1 (beta version), the CCI tool for ICD-10-CM was expanded
to identify four types of conditions:
\itemize{
\item \strong{Acute}: Examples include aortic embolism, bacterial infection, pregnancy,
and an initial encounter for an injury
\item \strong{Chronic}: Examples include malignant cancer, diabetes, obesity,
hypertension, and many mental health conditions
\item \strong{Both}: Examples include persistent asthma with (acute) exacerbation,
acute on chronic heart failure, and kidney transplant rejection
\item \strong{Not Applicable}: Examples include external cause of morbidity codes,
injury sequela codes, and codes starting with the letter Z for screening
or observation
}

In previous version (using ICD-9-CM), CCI classified conditions as chronic
or \strong{NonChronic}, so caution should be used when appying the CCI on data
using both ICD-9 and ICD-10.
}
}
\source{
\url{https://www.hcup-us.ahrq.gov/toolssoftware/chronic_icd10/CCI-ICD10CM-v2021-1.zip}
}
\usage{
CCI_icd10
}
\description{
Chronic Condition Indicator (CCI) for ICD-10-CM (beta version)
}
\details{
The version of CCI used in this package is \strong{beta version v2021.1}; The source documentation
\href{https://www.hcup-us.ahrq.gov/toolssoftware/chronic_icd10/chronic_icd10.jsp}{can be found here}
}
\keyword{datasets}
