% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CCI_icd9}
\alias{CCI_icd9}
\title{CCI for ICD-9-CM}
\format{
\describe{
\item{I9_DX}{ICD-9-CM diagnosis code (as char), without decimials, and
padded with white space, where applicable (see below)}
\item{I9_DX_desc}{Description of the ICD-9-CM diagnosis code}
\item{CCI}{Categorical variable inidcating if the ICD code represents an
\strong{chronic} or \strong{NonChronic} condition. This is unlike the ICD-10
beta version, so caution should be used when appying the CCI on data
using both ICD-9 and ICD-10}
\item{body_system}{Body system indicator of ICD-9 code, which is divided
18 categories}
ICD-9-CM codes are padded with whitespace, so shorter codes (\code{"0031 "})
will be treated like longer codes (\code{"00320"})
}
}
\source{
\url{https://www.hcup-us.ahrq.gov/toolssoftware/chronic/cci2015.csv}
}
\usage{
CCI_icd9
}
\description{
Chronic Condition Indicator (CCI) for ICD-9-CM
}
\details{
The version of CCI used in this package is \strong{2015 version}; The source documentation
\href{https://www.hcup-us.ahrq.gov/toolssoftware/chronic/chronic.jsp}{can be found here}
}
\keyword{datasets}
