% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CCS_labels}
\alias{CCS_labels}
\alias{CCS_dx9_label}
\alias{CCS_dx10_label}
\alias{CCS_pr9_label}
\alias{CCS_pr10_label}
\title{Look up the names of CCS categories from CCS codes.}
\format{
\describe{
\item{CCS}{Vector of CCS categories}
\item{CCS_label}{Text description of the CCS category}
\item{CCS_Level}{A column indicating what level the code belongs
to. If \code{main}, then it's a CCS code for single-level CCS. Otherwise,
it will be \code{lvl(n)}, where n is the level in the multi-level schema}
}
}
\usage{
CCS_labels

CCS_dx9_label

CCS_dx10_label

CCS_pr9_label

CCS_pr10_label
}
\description{
A dataset extracted from the  Healthcare Cost and Utilization
Project's (HCUP) Clinical Classifications Software (CCS)
}
\keyword{datasets}
