% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CCSR_DX_categories}
\alias{CCSR_DX_categories}
\title{Categories for CCSR diagnosis groups (ICD-10-CM)}
\format{
\describe{
\item{CCSR}{CCSR category}
\item{CCSR_desc}{Description of CCSR category}
\item{First3}{First 3 letters of CCSR category}
\item{I10DX_Chapter}{ICD-10-CM Diagnosis Chapter corresponding to
3-character abbreviation}
...
}
}
\source{
\url{https://www.hcup-us.ahrq.gov/toolssoftware/ccsr/DXCCSR_v2022-1.zip}
}
\usage{
CCSR_DX_categories
}
\description{
Categories for CCSR diagnosis groups (ICD-10-CM)
}
\details{
The version of CCSR used in this package is \strong{v2022.1}; The source documentation
\href{https://www.hcup-us.ahrq.gov/toolssoftware/ccsr/dxccsr.jsp}{can be found here}
}
\seealso{
See \code{\link{CCSR_labels}} for the labels of diagnosis codes and
procedure codes in one dataset
}
\keyword{datasets}
