% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CCSR_DX_mapping}
\alias{CCSR_DX_mapping}
\title{Mapping ICD-10-CM (diagnosis) codes to CCSR categories.}
\format{
\describe{
\item{I10_DX}{ICD-10-CM diagnosis code, without decimials}
\item{I10DX_text}{Description of the ICD-10-CM diagnosis code}
\item{default_CCSR_IP}{The default CCSR category for the principal
diagnosis for inpatient data}
\item{default_CCSR_OP}{The default CCSR category for first-listed
diagnosis for outpatient data}
\item{CCSR1}{}
\item{CCSR2}{}
\item{CCSR3}{}
\item{CCSR4}{}
\item{CCSR5}{}
\item{CCSR6}{CCSR categories assigned to a given ICD code. Some
codes have multiple categories. See example below}
The ICD-10-CM diagnosis code I11.9, Hypertensive heart disease without
heart failure, maps to a single CCSR category (CIR008 – Hypertension
with complications and secondary hypertension), whereas ICD-10-CM
diagnosis code I11.0, Hypertensive heart disease with heart failure,
maps to two CCSR categories (CIR008 – Hypertension with complications
and secondary hypertension and CIR019 – Heart failure)
}
}
\source{
\url{https://www.hcup-us.ahrq.gov/toolssoftware/ccsr/DXCCSR_v2022-1.zip}
}
\usage{
CCSR_DX_mapping
}
\description{
A dataset extracted from the  Healthcare Cost and Utilization
Project's (HCUP) Clinical Classifications Software Refined (CCSR)
for ICD-10-CM Diagnoses.
}
\details{
The overarching goal of the CCSR is to categorize codes into a manageable
number of clinically meaningful categories. The categories themselves should
capture a clinical concept, and the codes within a specific category should
maintain the clinical intent of that category. The version of CCSR used in
this package is \strong{v2022.1}; The source documentation
\href{https://www.hcup-us.ahrq.gov/toolssoftware/ccsr/dxccsr.jsp}{can be found here}
}
\keyword{datasets}
