% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CCSR_labels}
\alias{CCSR_labels}
\title{CCSR descriptions (both DX & PR)}
\format{
\describe{
\item{CCSR}{CCSR category}
\item{CCSR_desc}{Description of CCSR category}
\item{version}{Either \code{pr10} (for procedures) or \code{dx10} (for
diagnoses)}
}
}
\usage{
CCSR_labels
}
\description{
Descriptions of the CCSR categories for both diagnosis and
procedure CCSR groups
}
\seealso{
See \code{\link{CCSR_DX_categories}} and \code{\link{CCSR_PR_categories}}
for the two datasets used to generate this one.
}
\keyword{datasets}
