% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CCSR_PR_categories}
\alias{CCSR_PR_categories}
\title{Categories for CCSR procedure groups (ICD-10-PCS)}
\format{
\describe{
\item{CCSR}{CCSR category}
\item{CCSR_desc}{Description of CCSR category}
\item{clinical_domain}{Clinical domain of CCSR category}
\item{PCS_roots}{ICD-10-PCS Root Operations}
\item{PCS_bodyparts}{ICD-10-PCS Body Parts}
\item{PCS_devices}{ICD-10-PCS Devices}
\item{PCS_approaches}{ICD-10-PCS Approaches}
Columns starting with "PCS" contain comma separated values
provided by the HCUP
}
}
\source{
\url{https://www.hcup-us.ahrq.gov/toolssoftware/ccsr/PRCCSR_v2022-1.zip}
}
\usage{
CCSR_PR_categories
}
\description{
Categories for CCSR procedure groups (ICD-10-PCS)
}
\details{
The version of CCSR used in this package is \strong{v2022.1}. The source documentation
\href{https://www.hcup-us.ahrq.gov/toolssoftware/ccsr/prccsr.jsp}{can be found here}
}
\keyword{datasets}
