% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CCSR_PR_mapping}
\alias{CCSR_PR_mapping}
\title{Mapping ICD-10-PCS (procedure) codes to CCSR categories.}
\format{
\describe{
\item{I10_PR}{ICD-10-PCS procedure code, without decimials}
\item{I10_PR_desc}{Description of the ICD-10-CM procedure code}
\item{CCSR}{CCSR category assigned to the ICD code}
\item{CCSR_desc}{Description of the CCSR category}
}
}
\source{
\url{https://www.hcup-us.ahrq.gov/toolssoftware/ccsr/PRCCSR_v2022-1.zip}
}
\usage{
CCSR_PR_mapping
}
\description{
A dataset extracted from the  Healthcare Cost and Utilization
Project's (HCUP) Clinical Classifications Software Refined (CCSR)
for ICD-10-PCS Procedures.
}
\details{
The version of CCSR used in this package is \strong{v2022.1}. The source documentation
\href{https://www.hcup-us.ahrq.gov/toolssoftware/ccsr/prccsr.jsp}{can be found here}
}
\keyword{datasets}
