% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcup.data-package.R
\docType{package}
\name{hcup.data-package}
\alias{hcup.data}
\alias{hcup.data-package}
\title{hcup.data: Datasets Derived From HCUP Software Tools}
\description{
This package provides datasets used in the Healthcare Cost and Utilization Project (HCUP) Software tools
}
\author{
\strong{Maintainer}: Hunter Ratliff \email{hunterratliff.cv@gmail.com} (\href{https://orcid.org/0000-0003-4978-1998}{ORCID})

}
\keyword{internal}
