% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{proc_class_icd10}
\alias{proc_class_icd10}
\title{Procedure Classes Refined for the ICD-10-PCS}
\format{
\describe{
\item{I10_PR}{ICD-10-PCS procedure code}
\item{I10_PR_desc}{Description of the ICD-10-PCS procedure code}
\item{proc_class}{Categorical variable inidcating Procedure Classes Refined (see below)}
ICD-10-PCS procedure codes to one of four categories:
\itemize{
\item \strong{Minor Diagnostic}: Nonoperating room procedures that are diagnostic
(e.g. B244ZZZ, Ultrasonography of Right Heart)
\item \strong{Minor Therapeutic}: Nonoperating room procedures that are therapeutic
(e.g. 02HQ33Z, Insertion of Infusion Device into Right Pulmonary Artery,
Percutaneous Approach)
\item \strong{Major Diagnostic}: Procedures that are considered operating room procedures
that are performed for diagnostic reasons (e.g. 02BV0ZX, Excision of Superior
Vena Cava, Open Approach, Diagnostic)
\item \strong{Major Therapeutic}: Procedures that are considered operating room procedures
that are performed for therapeutic reasons (e.g. 0210093, Bypass Coronary Artery,
One Site from Coronary Artery with Autologous Venous Tissue, Open Approach).
}
}
}
\source{
\url{https://www.hcup-us.ahrq.gov/toolssoftware/procedureicd10/PClassR_v2022-2.zip}
}
\usage{
proc_class_icd10
}
\description{
Procedure Classes Refined for the ICD-10-PCS
}
\details{
The version of Procedure Classes Refined used in this package is version \strong{v2022.2};
The source documentation
\href{https://www.hcup-us.ahrq.gov/toolssoftware/procedureicd10/procedure_icd10.jsp}{can be found here}
}
\keyword{datasets}
