% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{proc_class_icd9}
\alias{proc_class_icd9}
\title{Procedure Classes for the ICD-9-CM procedure codes}
\format{
\describe{
\item{I9_PR}{ICD-9-CM procedure code, without decimials, left justified}
\item{I9_PR_desc}{Description of the ICD-9-CM procedure code}
\item{proc_class}{Categorical variable inidcating Procedure Classes (see below)}
ICD-9-CM procedure codes to one of four categories:
\itemize{
\item \strong{Minor Diagnostic}: Non-operating room procedures that are diagnostic (e.g., 87.03
CT scan of head)
\item \strong{Minor Therapeutic}: Non-operating room procedures that are therapeutic (e.g.,
02.41 Irrigate ventricular shunt)
\item \strong{Major Diagnostic}: All procedures considered valid operating room procedures by
the Diagnosis Related Group (DRG) grouper and that are performed for diagnostic reasons
(e.g., 01.14 Open brain biopsy)
\item \strong{Major Therapeutic}: All procedures considered valid operating room procedures by
the Diagnosis Related Group (DRG) grouper and that are performed for therapeutic reasons
(e.g., 39.24 Aorta-renal bypass).
}
}
}
\source{
\url{https://www.hcup-us.ahrq.gov/toolssoftware/procedure/pc2015.csv}
}
\usage{
proc_class_icd9
}
\description{
Procedure Classes for the ICD-9-CM procedure codes
}
\details{
The version of Procedure Classes used in this package is \strong{2015 version};
The source documentation
\href{https://www.hcup-us.ahrq.gov/toolssoftware/procedure/procedure.jsp}{can be found here}
}
\keyword{datasets}
