% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{valid_codes}
\alias{valid_codes}
\title{List of ICD codes}
\format{
List of four items
\describe{
\item{dx9}{Vector ICD-9-CM diagnosis codes}
\item{pr9}{Vector ICD-9-CM procedure codes}
\item{dx10}{Vector ICD-10-CM diagnosis codes}
\item{pr10}{Vector ICD-10-PCS procedure codes}
}
}
\usage{
valid_codes
}
\description{
A list of valid ICD codes, derived from the datasets elsewhere
in this package
}
\keyword{datasets}
